chore: add poetry package files
I'm going to managed this as a poetry project, and I'm going to call it
the "Excursion CLI", since an excursion is a synonym for "Drive", and
this is a fork of the "Drive CLI".

This is obviously dumb wordplay and I'm extremely open to changing this
should a better name arise or there becomes some conflict or confusion
wiht a better-known package.