chore: add pre-commit dev package for commit hooks
I want some linting to run before every commit, and the pre-commit
package is a great way to enforce that.

In particular, I want linting rules to be enforced, and I want commit
message conventions to be enforced.  This will be a forward-moving
effort; past commits don't need to be revisited.  I'm happy to preserved
the project's past commit history.

History revisions for pre-fork commits can be revisited if some
automation or tooling comes to depend upon the conventions established
here.